backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-dotenv==1.2.2
python-multipart
SQLAlchemy==2.0.39
psycopg[binary]==3.2.6
psycopg2-binary==2.9.10
livekit-api==0.8.2
stripe==12.0.0
